feat: token authentication

+++ src/main/kotlin/com/teamsparta/todo/domain/todo/service/TodoServiceImpl.kt
@@ -12,9 +12,11 @@ import com.teamsparta.todo.domain.todo.model.Todo
 import com.teamsparta.todo.domain.todo.model.toResponseDto
 import com.teamsparta.todo.domain.todo.model.toWithCommentsResponseDto
 import com.teamsparta.todo.domain.todo.repository.TodoRepository
+import com.teamsparta.todo.domain.user.repository.UserRepository
 import org.springframework.data.domain.PageRequest
 import org.springframework.data.domain.Sort
 import org.springframework.data.repository.findByIdOrNull
+import org.springframework.security.core.userdetails.User
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
@@ -22,6 +24,7 @@ import org.springframework.transaction.annotation.Transactional
 class TodoServiceImpl(
     private val todoRepository: TodoRepository,
     private val commentRepository: CommentRepository,
+    private val userRepository: UserRepository,
 ) : TodoService {
 
     override fun getTodoList(
@@ -34,7 +37,12 @@ class TodoServiceImpl(
         val todos = when {
             cursor == 0L -> {
                 if (writer == null) todoRepository.findAll(pageable)
-                else todoRepository.findAllByWriter(writer, pageable)
+                else {
+                    todoRepository.findAllByWriter(
+                        writer,
+                        pageable,
+                    )
+                }
             }
 
             sortDirection == Sort.Direction.DESC -> {
@@ -74,36 +82,54 @@ class TodoServiceImpl(
     }
 
     @Transactional
-    override fun createTodo(createTodoRequest: CreateTodoRequestDto): TodoResponseDto {
-        val (title, writer, content) = createTodoRequest
+    override fun createTodo(
+        user: User,
+        createTodoRequest: CreateTodoRequestDto,
+    ): TodoResponseDto {
+        val (title, content) = createTodoRequest
+        val writer = user.username
 
-        val todo = Todo(title = title, writer = writer, content = content)
+        if (userRepository.existsByUsername(writer) == false) throw IllegalArgumentException(
+            "Username: $writer does not exist.",
+        )
+        val todo =
+            Todo(title = title, writer = writer, content = content)
         return todoRepository.save(todo).toResponseDto()
     }
 
     @Transactional
     override fun updateTodo(
+        user: User,
         todoId: Long,
         updateTodoRequest: UpdateTodoRequestDto,
     ): TodoResponseDto {
         val todo =
             todoRepository.findByIdOrNull(todoId)
                 ?: throw ModelNotFoundException("Todo", todoId)
+        if (todo.writer != user.username) throw IllegalStateException(
+            "Username: ${user.username} is not the writer of todo (id: ${todoId}).",
+        )
 
-        val (title, writer, content) = updateTodoRequest
-        todo.update(title, writer, content)
+        val (title, content) = updateTodoRequest
+        todo.update(title, content)
 
         return todoRepository.save(todo).toResponseDto()
     }
 
     @Transactional
     override fun updateTodoStatus(
+        user: User,
         todoId: Long,
         updateTodoStatusRequest: UpdateTodoStatusRequestDto,
     ): TodoResponseDto {
         val todo =
             todoRepository.findByIdOrNull(todoId)
                 ?: throw ModelNotFoundException("Todo", todoId)
+
+        if (todo.writer != user.username) throw IllegalStateException(
+            "Username: ${user.username} is not the writer of the todo: ${todoId}.",
+        )
+
         val (status) = updateTodoStatusRequest
 
         todo.updateStatus(status)
@@ -112,10 +138,14 @@ class TodoServiceImpl(
     }
 
     @Transactional
-    override fun deleteTodo(todoId: Long) {
+    override fun deleteTodo(user: User, todoId: Long) {
         val todo =
             todoRepository.findByIdOrNull(todoId)
                 ?: throw ModelNotFoundException("Todo", todoId)
+        if (todo.writer != user.username) throw IllegalStateException(
+            "Username: ${user.username} is not the writer of the todo: ${todoId}.",
+        )
+
         val comments = commentRepository.findAllByTodoId(todoId)
         comments.forEach { commentRepository.delete(it) }
         todoRepository.delete(todo)
